Indexing test reorganized
