+++ tests/python/unittest/test_numpy_ndarray.py
@@ -377,93 +377,14 @@ def test_np_ndarray_copy():
 @with_seed()
 @use_np
 def test_np_ndarray_indexing():
-    def test_getitem(np_array, index):
-        """`is_scalar` indicates whether we should expect a scalar for the result.
-        If so, the indexed array of NDArray should call asscalar to compare
-        with numpy's indexed array."""
-        np_index = index
-        if isinstance(index, np.ndarray):
-            np_index = index.asnumpy()
-        if isinstance(index, tuple):
-            np_index = []
-            for idx in index:
-                if isinstance(idx, np.ndarray):
-                    np_index.append(idx.asnumpy())
-                else:
-                    np_index.append(idx)
-            np_index = tuple(np_index)
-
-        np_indexed_array = np_array[np_index]
-        mx_array = np.array(np_array, dtype=np_array.dtype)
-        mx_indexed_array = mx_array[index].asnumpy()
-        assert same(np_indexed_array, mx_indexed_array), 'Failed with index=%s' % str(index)
-
-    def test_setitem(np_array, index):
-        def assert_same(np_array, np_index, mx_array, mx_index, mx_value, np_value=None):
-            if np_value is not None:
-                np_array[np_index] = np_value
-            elif isinstance(mx_value, np.ndarray):
-                np_array[np_index] = mx_value.asnumpy()
-            else:
-                np_array[np_index] = mx_value
-            mx_array[mx_index] = mx_value
-            assert same(np_array, mx_array.asnumpy())
-
-        np_index = index
-        if isinstance(index, np.ndarray):
-            np_index = index.asnumpy()
-        if isinstance(index, tuple):
-            np_index = []
-            for idx in index:
-                if isinstance(idx, np.ndarray):
-                    np_index.append(idx.asnumpy())
-                else:
-                    np_index.append(idx)
-            np_index = tuple(np_index)
-
-        mx_array = np.array(np_array, dtype=np_array.dtype)
-        np_array = mx_array.asnumpy()
-        indexed_array_shape = np_array[np_index].shape
-        np_indexed_array = _np.random.randint(low=-10000, high=0, size=indexed_array_shape)
-        # test value is a numpy array without broadcast
-        assert_same(np_array, np_index, mx_array, index, np_indexed_array)
-        # test value is an numeric_type
-        assert_same(np_array, np_index, mx_array, index, _np.random.randint(low=-10000, high=0))
-        if len(indexed_array_shape) > 1:
-            # test ndarray with broadcast
-            assert_same(np_array, np_index, mx_array, index,
-                        np.random.uniform(low=-10000, high=0, size=(indexed_array_shape[-1],)))
-            # test numpy array with broadcast
-            assert_same(np_array, np_index, mx_array, index,
-                        _np.random.randint(low=-10000, high=0, size=(indexed_array_shape[-1],)))
-            # test list with broadcast
-            assert_same(np_array, np_index, mx_array, index,
-                        [_np.random.randint(low=-10000, high=0)] * indexed_array_shape[-1])
-
-    def test_getitem_autograd(np_array, index):
-        x = np.array(np_array, dtype=np_array.dtype)
-        x.attach_grad()
-        with autograd.record():
-            y = x[index]
-        y.backward()
-        value = np.ones_like(y)
-        x_grad = np.zeros_like(x)
-        x_grad[index] = value
-        assert same(x_grad.asnumpy(), x.grad.asnumpy())
-
-    def test_setitem_autograd(np_array, index):
-        x = np.array(np_array, dtype=np_array.dtype)
-        out_shape = x[index].shape
-        y = np.random.uniform(size=out_shape)
-        y.attach_grad()
-        try:
-            with autograd.record():
-                x[index] = y
-                assert False  # should not reach here
-        except mx.base.MXNetError as err:
-            assert str(err).find('Inplace operations (+=, -=, x[:]=, etc) are not supported when recording with') != -1
-
-    def np_int(index, int_type=_np.int32):
+    """
+    Test all indexing.
+    """
+    def np_int(index, int_type=np.int32):
+        """
+        Helper function for testing indexing that converts slices to slices of ints or None, and tuples to 
+        tuples of ints or None.
+        """
         def convert(num):
             if num is None:
                 return num
@@ -483,167 +404,262 @@ def test_np_ndarray_indexing():
         else:
             assert False
 
+    # Copied from test_ndarray.py. Under construction. 
+    def test_getitem(np_array, index, is_scalar=False):
+        """
+        `is_scalar` indicates whether we should expect a scalar for the result.
+        If so, the indexed array should call asscalar to compare
+        with numpy's indexed array.
+        np_array is a native numpy array.
+        """
+        np_index = index
+        if type(index) == mx.nd.NDArray:  # use of NDArray is prohibited
+            assert False
+        if isinstance(index, np.ndarray):
+            np_index = index.asnumpy()
+        if isinstance(index, tuple):
+            np_index = tuple([
+                idx.asnumpy() if isinstance(idx, mx.nd.NDArray) else idx
+                for idx in index]
+            )
+        np_indexed_array = np_array[np_index]
+        mx_np_array = np.array(np_array, dtype=np_array.dtype)
+        try:
+            mx_indexed_array = mx_np_array[index]
+        except Exception as e:
+            print('Failed with index = {}'.format(index))
+            raise e
+        if not is_scalar:
+            mx_indexed_array = mx_indexed_array.asnumpy()
+
+        assert same(np_indexed_array, mx_indexed_array), 'Failed with index = {}'.format(index)
+
+    def test_setitem(np_array, index, is_scalar=False):
+        """
+        `is_scalar` indicates whether we should expect a scalar for the result.
+        If so, the indexed array should call asscalar to compare
+        with numpy's indexed array.
+        np_array is a native numpy array.
+        """
+        def assert_same(np_array, np_index, mx_array, mx_index, mx_value, np_value=None):
+            if np_value is not None:
+                np_array[np_index] = np_value
+            elif isinstance(mx_value, np.ndarray):
+                np_array[np_index] = mx_value.asnumpy()
+            else:
+                np_array[np_index] = mx_value
+
+            try:
+                mx_array[mx_index] = mx_value
+            except Exception as e:
+                print('Failed with index = {}, value.shape = {}'.format(mx_index, mx_value.shape))
+                raise e
+
+            assert same(np_array, mx_array.asnumpy())
+
+        np_index = index  # keep this native numpy type
+        if isinstance(index, np.ndarray):
+            np_index = index.asnumpy()
+        if isinstance(index, tuple):
+            np_index = []
+            for idx in index:
+                if isinstance(idx, np.ndarray):
+                    np_index.append(idx.asnumpy())
+                else:
+                    np_index.append(idx)
+            np_index = tuple(np_index)
+
+        mx_array = np.array(np_array, dtype=np_array.dtype)  # mxnet.np.ndarray
+        np_array = mx_array.asnumpy()  # native numpy array
+        if is_scalar:
+            # test value is a numeric type
+            assert_same(np_array, np_index, mx_array, index, _np.random.randint(low=-10000, high=0))
+            value_nd = [_np.random.randint(low=-10000, high=0)]
+            assert_same(np_array, np_index, mx_array, index, value_nd, value_nd[0])
+        else:
+            indexed_array_shape = np_array[np_index].shape
+            np_indexed_array = _np.random.randint(low=-10000, high=0, size=indexed_array_shape)
+            # test value is a native numpy array without broadcast
+            assert_same(np_array, np_index, mx_array, index, np_indexed_array)
+            # test value is a mxnet numpy array without broadcast
+            assert_same(np_array, np_index, mx_array, index, np.array(np_indexed_array))
+            # test value is an numeric_type
+            assert_same(np_array, np_index, mx_array, index, _np.random.randint(low=-10000, high=0))
+            if len(indexed_array_shape) > 1:
+                np_value = _np.random.randint(low=-10000, high=0, size=(indexed_array_shape[-1],))
+                # test mxnet ndarray with broadcast
+                assert_same(np_array, np_index, mx_array, index, np.array(np_value))
+                # test native numpy array with broadcast
+                assert_same(np_array, np_index, mx_array, index, np_value)
+                # test list with broadcast
+                assert_same(np_array, np_index, mx_array, index,
+                            [_np.random.randint(low=-10000, high=0)] * indexed_array_shape[-1])
+
+    def test_getitem_autograd(np_array, index):
+        """
+        np_array: native numpy array.
+        """
+        x = np.array(np_array, dtype=np_array.dtype)
+        x.attach_grad()
+        with mx.autograd.record():
+            y = x[index]
+        if not isinstance(y, np.ndarray):
+            return
+        y.backward()
+        value = np.ones_like(y)
+        x_grad = np.zeros_like(x)
+        x_grad[index] = value
+        assert same(x_grad.asnumpy(), x.grad.asnumpy())
+
+    def test_setitem_autograd(np_array, index):
+        x = np.array(np_array, dtype=np_array.dtype)
+        if not isinstance(x[index], np.ndarray):
+            return  # x[index] is scalar
+        out_shape = x[index].shape
+        y = np.array(_np.random.uniform(size=out_shape))
+        y.attach_grad()
+        try:
+            with mx.autograd.record():
+                x[index] = y
+                # `a[None] = v` is equivalent to `a[...] = v` which doesn't raise
+                if index is not None:
+                    assert False, 'failed with index = {}'.format(index)  # should not reach here
+        except mx.base.MXNetError as err:
+            assert str(err).find('Inplace operations (+=, -=, x[:]=, etc) are not supported when recording with') != -1
+
     shape = (8, 16, 9, 9)
-    np_array = _np.arange(_np.prod(shape), dtype='int32').reshape(shape)
+    np_array = _np.arange(_np.prod(_np.array(shape)), dtype='int32').reshape(shape)  # native np array
     index_list = [
-        (),
-        0,
-        _np.int32(0),
-        _np.int64(0),
-        5,
-        _np.int32(5),
-        _np.int64(5),
-        -1,
-        _np.int32(-1),
-        _np.int64(-1),
-        slice(5),
-        np_int(slice(5), _np.int32),
-        np_int(slice(5), _np.int64),
-        slice(1, 5),
-        np_int(slice(1, 5), _np.int32),
-        np_int(slice(1, 5), _np.int64),
-        slice(1, 5, 2),
-        np_int(slice(1, 5, 2), _np.int32),
-        np_int(slice(1, 5, 2), _np.int64),
-        slice(7, 0, -1),
-        np_int(slice(7, 0, -1)),
-        np_int(slice(7, 0, -1), _np.int64),
-        slice(None, 6),
-        np_int(slice(None, 6)),
-        np_int(slice(None, 6), _np.int64),
-        slice(None, 6, 3),
-        np_int(slice(None, 6, 3)),
-        np_int(slice(None, 6, 3), _np.int64),
-        slice(1, None),
-        np_int(slice(1, None)),
-        np_int(slice(1, None), _np.int64),
-        slice(1, None, 3),
-        np_int(slice(1, None, 3)),
-        np_int(slice(1, None, 3), _np.int64),
-        slice(None, None, 2),
-        np_int(slice(None, None, 2)),
-        np_int(slice(None, None, 2), _np.int64),
-        slice(None, None, -1),
-        np_int(slice(None, None, -1)),
-        np_int(slice(None, None, -1), _np.int64),
-        slice(None, None, -2),
-        np_int(slice(None, None, -2), _np.int32),
-        np_int(slice(None, None, -2), _np.int64),
-        (slice(None), slice(None), 1, 8),
-        (slice(None), slice(None), -1, 8),
-        (slice(None), slice(None), 1, -8),
-        (slice(None), slice(None), -1, -8),
-        np_int((slice(None), slice(None), 1, 8)),
-        np_int((slice(None), slice(None), 1, 8), _np.int64),
-        (slice(None), slice(None), 1, 8),
-        np_int((slice(None), slice(None), -1, -8)),
-        np_int((slice(None), slice(None), -1, -8), _np.int64),
-        (slice(None), 2, slice(1, 5), 1),
-        np_int((slice(None), 2, slice(1, 5), 1)),
-        np_int((slice(None), 2, slice(1, 5), 1), _np.int64),
-        (1, 2, 3),
-        np_int((1, 2, 3)),
-        np_int((1, 2, 3), _np.int64),
-        (-1, -2, -3),
-        np_int((-1, -2, -3)),
-        np_int((-1, -2, -3), _np.int64),
-        (1, 2, 3, 4),
-        np_int((1, 2, 3, 4)),
-        np_int((1, 2, 3, 4), _np.int64),
-        (-4, -3, -2, -1),
-        np_int((-4, -3, -2, -1)),
-        np_int((-4, -3, -2, -1), _np.int64),
-        (slice(None, None, -1), 2, slice(1, 5), 1),
-        np_int((slice(None, None, -1), 2, slice(1, 5), 1)),
-        np_int((slice(None, None, -1), 2, slice(1, 5), 1), _np.int64),
-        (slice(None, None, -1), 2, slice(1, 7, 2), 1),
-        np_int((slice(None, None, -1), 2, slice(1, 7, 2), 1)),
-        np_int((slice(None, None, -1), 2, slice(1, 7, 2), 1), _np.int64),
-        (slice(1, 8, 2), slice(14, 2, -2), slice(3, 8), slice(0, 7, 3)),
-        np_int((slice(1, 8, 2), slice(14, 2, -2), slice(3, 8), slice(0, 7, 3))),
-        np_int((slice(1, 8, 2), slice(14, 2, -2), slice(3, 8), slice(0, 7, 3)), _np.int64),
-        (slice(1, 8, 2), 1, slice(3, 8), 2),
-        np_int((slice(1, 8, 2), 1, slice(3, 8), 2)),
-        np_int((slice(1, 8, 2), 1, slice(3, 8), 2), _np.int64),
-        [1],
-        [1, 2],
-        [2, 1, 3],
-        [7, 5, 0, 3, 6, 2, 1],
-        _np.array([6, 3], dtype=_np.int32),
-        _np.array([[3, 4], [0, 6]], dtype=_np.int32),
-        _np.array([[7, 3], [2, 6], [0, 5], [4, 1]], dtype=_np.int32),
-        _np.array([[7, 3], [2, 6], [0, 5], [4, 1]], dtype=_np.int64),
-        _np.array([[2], [0], [1]], dtype=_np.int32),
-        _np.array([[2], [0], [1]], dtype=_np.int64),
-        np.array([4, 7], dtype=_np.int32),
-        np.array([4, 7], dtype=_np.int64),
-        np.array([[3, 6], [2, 1]], dtype=_np.int32),
-        np.array([[3, 6], [2, 1]], dtype=_np.int64),
-        np.array([[7, 3], [2, 6], [0, 5], [4, 1]], dtype=_np.int32),
-        np.array([[7, 3], [2, 6], [0, 5], [4, 1]], dtype=_np.int64),
-        (1, [2, 3]),
-        (1, [2, 3], _np.array([[3], [0]], dtype=_np.int32)),
-        (1, [2, 3]),
-        (1, [2, 3], _np.array([[3], [0]], dtype=_np.int64)),
-        (1, [2], _np.array([[5], [3]], dtype=_np.int32), slice(None)),
-        (1, [2], _np.array([[5], [3]], dtype=_np.int64), slice(None)),
-        (1, [2, 3], _np.array([[6], [0]], dtype=_np.int32), slice(2, 5)),
-        (1, [2, 3], _np.array([[6], [0]], dtype=_np.int64), slice(2, 5)),
-        (1, [2, 3], _np.array([[4], [7]], dtype=_np.int32), slice(2, 5, 2)),
-        (1, [2, 3], _np.array([[4], [7]], dtype=_np.int64), slice(2, 5, 2)),
-        (1, [2], _np.array([[3]], dtype=_np.int32), slice(None, None, -1)),
-        (1, [2], _np.array([[3]], dtype=_np.int64), slice(None, None, -1)),
-        (1, [2], _np.array([[3]], dtype=_np.int32), np.array([[5, 7], [2, 4]], dtype=_np.int64)),
-        (1, [2], np.array([[4]], dtype=_np.int32), np.array([[1, 3], [5, 7]], dtype='int64')),
-        [0],
-        [0, 1],
-        [1, 2, 3],
-        [2, 0, 5, 6],
-        ([1, 1], [2, 3]),
-        ([1], [4], [5]),
-        ([1], [4], [5], [6]),
-        ([[1]], [[2]]),
-        ([[1]], [[2]], [[3]], [[4]]),
-        (slice(0, 2), [[1], [6]], slice(0, 2), slice(0, 5, 2)),
-        ([[[[1]]]], [[1]], slice(0, 3), [1, 5]),
-        ([[[[1]]]], 3, slice(0, 3), [1, 3]),
-        ([[[[1]]]], 3, slice(0, 3), 0),
-        ([[[[1]]]], [[2], [12]], slice(0, 3), slice(None)),
-        ([1, 2], slice(3, 5), [2, 3], [3, 4]),
-        ([1, 2], slice(3, 5), (2, 3), [3, 4]),
-        range(4),
-        range(3, 0, -1),
-        (range(4,), [1]),
-        # Test Ellipsis ('...')
-        ((1, Ellipsis, -1), False),
-        ((slice(2), Ellipsis, None, 0), False),
-        ((Ellipsis, 3, 2, slice(1, 2)), False)
-        # Test newaxis, basic indexing
-        (None, False),
-        ((1, None, -2, 3, -4), False),
-        ((1, slice(2, 5), None), False), 
-        ((slice(None), slice(1, 4), None, slice(2, 3)), False), 
-        ((slice(1, 3), slice(1, 3), slice(1, 3), slice(1, 3), None), False), 
-        ((slice(1, 3), slice(1, 3), None, slice(1, 3), slice(1, 3)), False), 
-        ((None, slice(1, 2), 3, None), False),
-        ((1, None, 2, 3, None, None, 4), False),
-        ((None, Ellipsis, 3, 2, slice(1, 2), None,), False),
-        # Test newaxis, advanced indexing
-        (([1, 2], slice(3, 5), None, None, [3, 4]), False),
-        ((slice(None), slice(3, 5), None, None, [2, 3], [3, 4]), False),
-        ((slice(None), slice(3, 5), None, [2, 3], None, [3, 4]), False),
-        ((None, slice(None), slice(3, 5), [2, 3], None, [3, 4]), False),
-        # slice(0, 0) does not support output zero-size tensor yet
+                # Basic indexing
+                # Single int as index
+                (0, False), (np.int32(0), False), (np.int64(0), False),
+                (5, False), (np.int32(5), False), (np.int64(5), False),
+                (-1, False), (np.int32(-1), False), (np.int64(-1), False),
+                # Slicing as index
+                (slice(5), False), (np_int(slice(5), np.int32), False), (np_int(slice(5), np.int64), False),
+                (slice(1, 5), False), (np_int(slice(1, 5), np.int32), False), (np_int(slice(1, 5), np.int64), False),
+                (slice(1, 5, 2), False), (np_int(slice(1, 5, 2), np.int32), False),
+                (np_int(slice(1, 5, 2), np.int64), False),
+                (slice(7, 0, -1), False), (np_int(slice(7, 0, -1)), False),
+                (np_int(slice(7, 0, -1), np.int64), False),
+                (slice(None, 6), False), (np_int(slice(None, 6)), False),
+                (np_int(slice(None, 6), np.int64), False),
+                (slice(None, 6, 3), False), (np_int(slice(None, 6, 3)), False),
+                (np_int(slice(None, 6, 3), np.int64), False),
+                (slice(1, None), False), (np_int(slice(1, None)), False),
+                (np_int(slice(1, None), np.int64), False),
+                (slice(1, None, 3), False), (np_int(slice(1, None, 3)), False),
+                (np_int(slice(1, None, 3), np.int64), False),
+                (slice(None, None, 2), False), (np_int(slice(None, None, 2)), False),
+                (np_int(slice(None, None, 2), np.int64), False),
+                (slice(None, None, -1), False),
+                (np_int(slice(None, None, -1)), False), (np_int(slice(None, None, -1), np.int64), False),
+                (slice(None, None, -2), False),
+                (np_int(slice(None, None, -2), np.int32), False), (np_int(slice(None, None, -2), np.int64), False),
+                # Multiple ints as indices
+                ((1, 2, 3), False),
+                (np_int((1, 2, 3)), False),
+                (np_int((1, 2, 3), np.int64), False),
+                ((-1, -2, -3), False),
+                (np_int((-1, -2, -3)), False),
+                (np_int((-1, -2, -3), np.int64), False),
+                ((1, 2, 3, 4), True),
+                (np_int((1, 2, 3, 4)), True),
+                (np_int((1, 2, 3, 4), np.int64), True),
+                ((-4, -3, -2, -1), True),
+                (np_int((-4, -3, -2, -1)), True),
+                (np_int((-4, -3, -2, -1), np.int64), True),
+                # slice(None) as indices
+                ((slice(None), slice(None), 1, 8), False),
+                ((slice(None), slice(None), -1, 8), False),
+                ((slice(None), slice(None), 1, -8), False),
+                ((slice(None), slice(None), -1, -8), False),
+                (np_int((slice(None), slice(None), 1, 8)), False),
+                (np_int((slice(None), slice(None), 1, 8), np.int64), False),
+                ((slice(None), slice(None), 1, 8), False),
+                (np_int((slice(None), slice(None), -1, -8)), False),
+                (np_int((slice(None), slice(None), -1, -8), np.int64), False),
+                ((slice(None), 2, slice(1, 5), 1), False),
+                (np_int((slice(None), 2, slice(1, 5), 1)), False),
+                (np_int((slice(None), 2, slice(1, 5), 1), np.int64), False),
+                # Mixture of ints and slices as indices
+                ((slice(None, None, -1), 2, slice(1, 5), 1), False),
+                (np_int((slice(None, None, -1), 2, slice(1, 5), 1)), False),
+                (np_int((slice(None, None, -1), 2, slice(1, 5), 1), np.int64), False),
+                ((slice(None, None, -1), 2, slice(1, 7, 2), 1), False),
+                (np_int((slice(None, None, -1), 2, slice(1, 7, 2), 1)), False),
+                (np_int((slice(None, None, -1), 2, slice(1, 7, 2), 1), np.int64), False),
+                ((slice(1, 8, 2), slice(14, 2, -2), slice(3, 8), slice(0, 7, 3)), False),
+                (np_int((slice(1, 8, 2), slice(14, 2, -2), slice(3, 8), slice(0, 7, 3))), False),
+                (np_int((slice(1, 8, 2), slice(14, 2, -2), slice(3, 8), slice(0, 7, 3)), np.int64), False),
+                ((slice(1, 8, 2), 1, slice(3, 8), 2), False),
+                (np_int((slice(1, 8, 2), 1, slice(3, 8), 2)), False),
+                (np_int((slice(1, 8, 2), 1, slice(3, 8), 2), np.int64), False),
+                # Test Ellipsis ('...')
+                ((1, Ellipsis, -1), False),
+                ((slice(2), Ellipsis, None, 0), False),
+                # Test newaxis
+                (None, False),
+                ((1, None, -2, 3, -4), False),
+                ((1, slice(2, 5), None), False), 
+                ((slice(None), slice(1, 4), None, slice(2, 3)), False), 
+                ((slice(1, 3), slice(1, 3), slice(1, 3), slice(1, 3), None), False), 
+                ((slice(1, 3), slice(1, 3), None, slice(1, 3), slice(1, 3)), False), 
+                ((None, slice(1, 2), 3, None), False),
+                ((1, None, 2, 3, None, None, 4), False),
+                # Advanced indexing
+                (([1, 2], slice(3, 5), None, None, [3, 4]), False),
+                ((slice(None), slice(3, 5), None, None, [2, 3], [3, 4]), False),
+                ((slice(None), slice(3, 5), None, [2, 3], None, [3, 4]), False),
+                ((None, slice(None), slice(3, 5), [2, 3], None, [3, 4]), False),
+                ([1], False), ([1, 2], False), ([2, 1, 3], False), ([7, 5, 0, 3, 6, 2, 1], False),
+                (np.array([6, 3], dtype=np.int32), False),
+                (np.array([[3, 4], [0, 6]], dtype=np.int32), False),
+                (np.array([[7, 3], [2, 6], [0, 5], [4, 1]], dtype=np.int32), False),
+                (np.array([[7, 3], [2, 6], [0, 5], [4, 1]], dtype=np.int64), False),
+                (np.array([[2], [0], [1]], dtype=np.int32), False),
+                (np.array([[2], [0], [1]], dtype=np.int64), False),
+                (np.array([4, 7], dtype=np.int32), False),
+                (np.array([4, 7], dtype=np.int64), False),
+                (np.array([[3, 6], [2, 1]], dtype=np.int32), False),
+                (np.array([[3, 6], [2, 1]], dtype=np.int64), False),
+                (np.array([[7, 3], [2, 6], [0, 5], [4, 1]], dtype=np.int32), False),
+                (np.array([[7, 3], [2, 6], [0, 5], [4, 1]], dtype=np.int64), False),
+                ((1, [2, 3]), False), ((1, [2, 3], np.array([[3], [0]], dtype=np.int32)), False),
+                ((1, [2, 3]), False), ((1, [2, 3], np.array([[3], [0]], dtype=np.int64)), False),
+                ((1, [2], np.array([[5], [3]], dtype=np.int32), slice(None)), False),
+                ((1, [2], np.array([[5], [3]], dtype=np.int64), slice(None)), False),
+                ((1, [2, 3], np.array([[6], [0]], dtype=np.int32), slice(2, 5)), False),
+                ((1, [2, 3], np.array([[6], [0]], dtype=np.int64), slice(2, 5)), False),
+                ((1, [2, 3], np.array([[4], [7]], dtype=np.int32), slice(2, 5, 2)), False),
+                ((1, [2, 3], np.array([[4], [7]], dtype=np.int64), slice(2, 5, 2)), False),
+                ((1, [2], np.array([[3]], dtype=np.int32), slice(None, None, -1)), False),
+                ((1, [2], np.array([[3]], dtype=np.int64), slice(None, None, -1)), False),
+                ((1, [2], np.array([[3]], dtype=np.int32), np.array([[5, 7], [2, 4]], dtype=np.int64)), False),
+                ((1, [2], np.array([[4]], dtype=np.int32), np.array([[1, 3], [5, 7]], dtype='int64')),
+                False),
+                ([0], False), ([0, 1], False), ([1, 2, 3], False), ([2, 0, 5, 6], False),
+                (([1, 1], [2, 3]), False), (([1], [4], [5]), False), (([1], [4], [5], [6]), False),
+                (([[1]], [[2]]), False), (([[1]], [[2]], [[3]], [[4]]), False),
+                ((slice(0, 2), [[1], [6]], slice(0, 2), slice(0, 5, 2)), False),
+                (([[[[1]]]], [[1]], slice(0, 3), [1, 5]), False),
+                (([[[[1]]]], 3, slice(0, 3), [1, 3]), False),
+                (([[[[1]]]], 3, slice(0, 3), 0), False),
+                (([[[[1]]]], [[2], [12]], slice(0, 3), slice(None)), False),
+                (([1, 2], slice(3, 5), [2, 3], [3, 4]), False),
+                (([1, 2], slice(3, 5), (2, 3), [3, 4]), False),
+                # Output zero size tensor not supported yet.
+                # ((slice(0, 0), slice(0, 0), 1, 2), False),
+                # ((slice(0, 0), slice(0, 0), slice(0, 0), slice(0, 0)), True),
     ]
     for index in index_list:
-        test_getitem(np_array, index)
-        test_setitem(np_array, index)
-        test_getitem_autograd(np_array, index)
+        test_getitem(np_array, index[0], index[1])
+        test_setitem(np_array, index[0], index[1])
+        test_getitem_autograd(np_array, index[0])
         if not isinstance(index, tuple) or len(index) != 0:
             # When index = (), this is same a[()] = b is equivalent to b.copyto(a)
             # which should have no problem to do autograd
-            test_setitem_autograd(np_array, index)
-
+            test_setitem_autograd(np_array, index[0])
 
 @with_seed()
 @use_np
